requirements.txt
google-api-python-client==2.165.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.4
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pdfplumber==0.11.5
pypdf==5.4.0
pytesseract==0.3.13
pdf2image==1.17.0
slack_sdk==3.35.0
SQLAlchemy==2.0.38
APScheduler==3.11.0
